pyproject.toml
[project]
name = "onchaindata"
version = "0.1.0"
description = "Capstone project package for FA DAE2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dlt>=1.17.0",
    "jupyter>=1.1.1",
    "pandas>=2.3.3",
    "polars>=1.33.1",
    "psycopg>=3.2.10",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.0.0",
    "rich>=14.1.0",
    "snowflake-connector-python>=3.17.4",
    "sqlalchemy>=2.0.43",
]

[tool.uv.sources]
onchaindata = { path = "./src" }

[tool.uv]
package = true
